Okay, it all functions.
To test it get to the drop down menu, and select the test section.  It
will automatically open up the new page which will just be the users
avatar, the country and state they live in.  That I believe is enough of
a demo but we can put together anything else we need later.